+++ BookGen.Launch/App.xaml.cs
@@ -3,9 +3,11 @@
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
+using BookGen.ShellHelper;
 using System;
 using System.Collections.Generic;
 using System.Windows;
+using System.Windows.Navigation;
 using System.Windows.Shell;
 
 namespace BookGen.Launch
@@ -35,6 +37,16 @@ namespace BookGen.Launch
                 list.JumpItems.Add(task);
             }
             list.Apply();
+        }
+
+        protected override void OnStartup(StartupEventArgs e)
+        {
+            base.OnStartup(e);
+            var rerminalInstall = TerminalProfileInstaller.TryInstall();
+            if (rerminalInstall == false)
+            {
+                MessageBox.Show(Launch.Properties.Resources.TerminalProfileInstallFail);
+            }
         }
     }
 }