+++ BookGen.Launch/Code/StartShellCommand.cs
@@ -1,9 +1,9 @@
 ﻿//-----------------------------------------------------------------------------
-// (c) 2021 Ruzsinszki Gábor
+// (c) 2021-2022 Ruzsinszki Gábor
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
-using BookGen.Resources;
+using BookGen.ShellHelper;
 using System;
 using System.IO;
 using System.Windows;
@@ -12,45 +12,12 @@ namespace BookGen.Launch.Code
 {
     internal sealed class StartShellCommand : ProcessCommandBase
     {
-        private bool TryCreateShellScript(string folder, out string shellScriptPath)
-        {
-            try
-            {
-#pragma warning disable S5445 // Insecure temporary file creation methods should not be used
-                string? name = Path.GetTempFileName();
-#pragma warning restore S5445 // Insecure temporary file creation methods should not be used
-                name = Path.ChangeExtension(name, ".ps1");
-
-                using (var file = File.CreateText(name))
-                {
-                    var dn = ResourceHandler.GetResourceFile<KnownFile>("Powershell/completer.dn.ps1");
-                    var bg = ResourceHandler.GetResourceFile<KnownFile>("Powershell/completer.ps1");
-                    var sh = ResourceHandler.GetResourceFile<KnownFile>("Powershell/shell.ps1");
-
-                    file.WriteLine("$env:Path += \";{0}\"", AppContext.BaseDirectory);
-                    file.WriteLine("Set-Location \"{0}\"", folder);
-                    file.WriteLine(dn);
-                    file.WriteLine(bg);
-                    file.WriteLine("Remove-Item \"{0}\"", name);
-                    file.WriteLine(sh);
-                }
-
-                shellScriptPath = name;
-                return true;
-            }
-            catch (IOException)
-            {
-                shellScriptPath = string.Empty;
-                return false;
-            }
-        }
-
         private bool Launch(string shellScript)
         {
-            if (isWindowsTerminalInstalled && Properties.Settings.Default.UseWindowsTerminal)
-                return RunProgram(WindowsTerminalExe, $"new-tab -p \"Powershell\" --title \"BookGen shell\" --colorScheme \"Campbell Powershell\" powershell.exe -ExecutionPolicy Bypass -NoExit -File \"{shellScript}\"");
-            else if (isPsCoreInstalled)
-                return RunProgram(PowershellCoreExe, $"-NoExit -File \"{shellScript}\"");
+            if (InstallStatus.IsWindowsTerminalInstalled && Properties.Settings.Default.UseWindowsTerminal)
+                return RunProgram(InstallDetector.WindowsTerminalExe, $"new-tab -p \"Powershell\" --title \"BookGen shell\" --colorScheme \"Campbell Powershell\" powershell.exe -ExecutionPolicy Bypass -NoExit -File \"{shellScript}\"");
+            else if (InstallStatus.IsPsCoreInstalled)
+                return RunProgram(InstallDetector.PowershellCoreExe, $"-NoExit -File \"{shellScript}\"");
             else
                 return RunProgram(PowershellExe, $"-ExecutionPolicy Bypass -NoExit -File \"{shellScript}\"");
         }
@@ -70,14 +37,14 @@ namespace BookGen.Launch.Code
                 return;
             }
 
-            string _shellScript;
-            if (!TryCreateShellScript(folder, out _shellScript))
+            string shellScript = Path.Combine(AppContext.BaseDirectory, "BookGenShell.ps1");
+            if (!File.Exists(shellScript))
             {
                 Message(Properties.Resources.ShellScriptWriteFail, MessageBoxImage.Error);
                 return;
             }
 
-            if (!Launch(_shellScript))
+            if (!Launch(shellScript))
             {
                 Message(Properties.Resources.ShellScriptStartFail, MessageBoxImage.Error);
             }