+++ BookGen/GeneratorSteps/CreatePrintableHtml.cs
@@ -43,18 +43,16 @@ namespace BookGen.GeneratorSteps
             foreach (var chapter in settings.TocContents.Chapters)
             {
                 log.Info("Processing: {0}...", chapter);
+                reindexer.AddMarkdown($"# {chapter}\r\n");
                 foreach (var file in settings.TocContents.GetLinksForChapter(chapter).Select(l => l.Url))
                 {
                     log.Detail("Processing file for print output: {0}", file);
                     var input = settings.SourceDirectory.Combine(file);
 
-                    var inputContent = input.ReadFile(log);
+                    var inputContent = input.ReadFile(log, true);
                     reindexer.AddMarkdown(inputContent);
                 }
             }
-
-            //FsPath? debug = settings.OutputDirectory.Combine("debug.md");
-            //debug.WriteFile(log, reindexer.ToString());
 
             Content.Content = pipeline.RenderMarkdown(reindexer.ToString());