+++ WpLoad/Commands/Up.cs
@@ -3,6 +3,7 @@
 // This code is licensed under MIT license (see LICENSE for details)
 //-----------------------------------------------------------------------------
 
+using BookGen.Gui.ArgumentParser;
 using System.Web;
 using WordPressPCL;
 using WordPressPCL.Models;
@@ -12,35 +13,18 @@ using WpLoad.Services;
 
 namespace WpLoad.Commands
 {
-    internal class Up : IAsyncCommand
+    internal class Up : LoadCommandBase
     {
-        public string CommandName => nameof(Up);
+        public override string CommandName => nameof(Up);
 
-        public async Task<ExitCode> Execute(ILog log, IReadOnlyList<string> arguments)
+        public override async Task<ExitCode> Execute(ILog log, IReadOnlyList<string> arguments)
         {
-            if (arguments.TryGetArgument(0, out string? site))
+            UpArguments args = new();
+            ArgumentParser.ParseArguments(arguments, args);
+
+            if (TryConfigureFolderAndClient(log, args, out var client))
             {
-                if (!arguments.TryGetArgument(1, out string? folder))
-                {
-                    folder = Environment.CurrentDirectory;
-                }
-
-                if (!Directory.Exists(folder))
-                {
-                    log.Error($"{folder} doesn't exist");
-                    return ExitCode.Fail;
-                }
-
-                log.Info("Configuring connection...");
-                if (!ClientService.TryConfifgureConnection(site, out WordPressClient? client))
-                {
-                    log.Error($"{site} profile doesn't exist");
-                    return ExitCode.Fail;
-                }
-
-
-                var mediaFiles = FileServices.GetSupportedFilesInDirectory(folder);
-
+                var mediaFiles = FileServices.GetSupportedFilesInDirectory(args.Path);
                 try
                 {
                     await UploadMedia(log, client, mediaFiles.mediaFiles);
@@ -53,7 +37,6 @@ namespace WpLoad.Commands
                     log.Error(ex);
                     return ExitCode.Fail;
                 }
-
             }
             return ExitCode.BadParameters;
         }
@@ -62,14 +45,14 @@ namespace WpLoad.Commands
         {
             ProgressReporter reporter = new ProgressReporter(log);
             reporter.Start();
-            foreach (var html in htmls)
+            await Parallel.ForEachAsync(htmls, async (html, ct) =>
             {
                 string content = await File.ReadAllTextAsync(html);
                 log.Info($"Uploading {Path.GetFileName(html)}...");
-                await client.Posts.Create(CreatePost(html, content));
+                await client.Posts.CreateAsync(CreatePost(html, content));
                 reporter.Report(html);
 
-            }
+            });
             reporter.Stop();
         }
 
@@ -77,14 +60,14 @@ namespace WpLoad.Commands
         {
             ProgressReporter reporter = new ProgressReporter(log);
             reporter.Start();
-            foreach (var media in mediaFiles)
+            await Parallel.ForEachAsync(mediaFiles, async (media, ct) =>
             {
                 log.Info($"Uploading {Path.GetFileName(media)}...");
-                await client.Media.Create(media,
-                                          HttpUtility.UrlEncode(Path.GetFileName(media)),
-                                          FileServices.GetMimeType(media));
+                await client.Media.CreateAsync(media,
+                          HttpUtility.UrlEncode(Path.GetFileName(media)),
+                          FileServices.GetMimeType(media));
                 reporter.Report(media);
-            }
+            });
             reporter.Stop();
         }